Renamed combiner/server.py to combiner/combiner.py
